refactor(files)!: Rename `gen-readme.py` to `main.py`
